Move the redux folder up to make a sibling of component
