+++ src/components/Categories.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { checkStatus } from './redux/categories/categories';
+import { checkStatus } from '../redux/categories/categories';
 
 export default function Categories() {
   const dispatch = useDispatch();
