+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import configureStore from './components/redux/configureStore';
+import configureStore from './redux/configureStore';
 import './index.css';
 import App from './components/App';
